added comments

+++ app/Http/Controllers/TaskController.php
@@ -8,9 +8,10 @@ use App\Models\Task;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Gate;
 
+# php artisan make:controller TaskController
 class TaskController extends Controller
 {
-    // Metoda do dodawania kategorii
+    # Method adds task to category
     public function store(StoreRequest $request, $category)
     {
         $task = Task::make([
@@ -30,12 +31,15 @@ class TaskController extends Controller
         return back()->withSuccess('Zadanie zostało dodane');
     }
 
+    # Method updates data in task
     public function update(StoreRequest $request, Task $task)
     {
         if (Gate::inspect('manage', $task)->allowed() === false) {
             return $this->unauthorized();
         }
 
+        # if ( [ now() > task->deadline ] or [ task is finished ] )
+        # => user cannot update data in task
         if ($task->can_manage === false) {
             return $this->cannotManage($task);
         }
@@ -51,6 +55,7 @@ class TaskController extends Controller
         return redirect()->route('category.show', $task->category)->withSuccess("Dane zostały zaktualizowane");
     }
 
+    # Method return view
     public function edit(Task $task)
     {
         if (Gate::inspect('manage', $task)->allowed() === false) {
@@ -66,16 +71,22 @@ class TaskController extends Controller
         ]);
     }
 
+    # Method deletes task
     public function delete(Task $task)
     {
         if (Gate::inspect('manage', $task)->allowed() === false) {
             return $this->unauthorized();
         }
 
+        if ($task->can_manage === false) {
+            return $this->cannotManage($task);
+        }
+
         $task->delete();
         return back()->withSuccess("Zadanie zostało usunięte");
     }
 
+    # Method completes tasks
     public function finish(Task $task)
     {
         if (Gate::inspect('manage', $task)->allowed() === false) {
@@ -90,16 +101,18 @@ class TaskController extends Controller
         return back()->withSuccess("Zadanie zostało ukończone");
     }
 
+    # Method redirect user to route('category.index')
+    # when user don't have permissions to task
     private function unauthorized()
     {
         return redirect()->route('category.index')->with('error', "Brak uprawnień do wykonania akcji");
     }
 
+    # Method redirect user to route('category.show') when he try do something on task
+    # WHEN [ now() > task->deadline ] or [ task is finished ]
+    # exclude method finish
     private function cannotManage($task)
     {
-        return redirect()->route('category.show', $task->category)->with('error', "Nie można edytować zadań po terminie ukończenia");
+        return redirect()->route('category.show', $task->category)->with('error', "Tego zadania nie można już edytować");
     }
 }
-
-# Zadań po terminiu nie można edytować oraz usuwać => tylko zakonczyć
-# Czy w update ma być możliwość zmiany terminu ?
